requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
ca-certificates=2021.1.19=haa95532_0
certifi=2020.12.5=py38haa95532_0
icu=58.2=ha925a31_3
jpeg=9b=hb83a4c4_2
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.14.0=h7755175_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
msys2-conda-epoch=20160418=1
openssl=1.1.1j=h2bbff1b_0
pip=21.0.1=py38haa95532_0
protobuf=3.14.0=py38h885f38d_1
pyqt=5.9.2=py38ha925a31_4
python=3.8.5=h5fd99cc_1
python_abi=3.8=1_cp38
qt=5.9.7=vc14h73c81de_0
rx=3.1.1=py_0
setuptools=52.0.0=py38haa95532_0
sip=4.19.13=py38ha925a31_0
six=1.15.0=pyh9f0ad1d_0
sqlite=3.33.0=h2a8f88b_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.36.2=pyhd3eb1b0_0
wincertstore=0.2=py38_0
zaber-motion=2.1.2=py38h2865ccd_0
zlib=1.2.11=h62dcd97_4
